Moved strategy object
